gst-nmos-rs: colocate session tests with mxl and udp modules
Move transport-specific #[cfg(test)] modules into session/mxl.rs
and session/udp.rs; keep shared activation, proto mapping, and
auto-activate coverage in session/mod.rs with a small support
fixture module.
